+++ src/main/java/com/goorm/tablepick/domain/reservation/service/ReservationService/ReservationServiceV2.java
@@ -51,7 +51,7 @@ public class ReservationServiceV2 {
 
         // 슬롯 카운트 검증
         Long count = reservationSlot.getCount();
-        Long maxCapacity = restaurant.getMaxCapacity();
+        int maxCapacity = restaurant.getMaxCapacity();
         if (count >= maxCapacity) {
             throw new ReservationException(ReservationErrorCode.EXCEED_RESERVATION_LIMIT);
         }
@@ -94,7 +94,7 @@ public class ReservationServiceV2 {
 
         // 슬롯 카운트 검증
         Long count = reservationSlot.getCount();
-        Long maxCapacity = restaurant.getMaxCapacity();
+        int maxCapacity = restaurant.getMaxCapacity();
         if (count >= maxCapacity) {
             throw new ReservationException(ReservationErrorCode.EXCEED_RESERVATION_LIMIT);
         }
